# api/docs/swagger.yaml
definitions:
...
  template_variables.FailureResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  var_api_gateway.EntityFilesSwag:
    properties:
      comment:
        type: string
    type: object
  var_api_gateway.FilesResponse:
    properties:
      file_path:
        type: string
      file_url:
        type: string
    type: object
...
info:
  contact: {}
  license: {}
paths:
  /v1/file-upload:
    post:
      consumes:
      - multipart/form-data
      - application/json
      description: API for creating file
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: use_p
        in: query
        name: use_p
        type: string
      - description: entity_id
        in: query
        name: entity_id
        type: string
      - description: property_id
        in: query
        name: property_id
        type: string
      - description: file_name_id
        in: query
        name: file_name_id
        type: string
      - description: region
        in: body
        name: region
        required: true
        schema:
          $ref: '#/definitions/var_api_gateway.EntityFilesSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/var_api_gateway.FilesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: Create file
      tags:
      - file_upload
  /v1/image-upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: image
        in: formData
        name: image
        required: true
        type: file
      - description: entity_id
        in: query
        name: entity_id
        type: string
      - description: file_name_id
        in: query
        name: file_name_id
        type: string
      - description: property_id
        in: query
        name: property_id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/var_api_gateway.FilesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      tags:
      - file_upload
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"